Correct imaginary number parsing

+++ src/datatype.cxx
@@ -3,6 +3,7 @@
 
 #include <limits>
 #include <regex>
+#include <sstream>
 #include <stdexcept>
 #include <string>
 
@@ -312,8 +313,8 @@ void yaml_decode_complex(const YAML::Node &node, complex<T> &val) {
   smatch m;
   bool didmatch = regex_match(str, m, cmplx);
   assert(didmatch);
-  const T re = m[1].matched ? static_cast<T>(stold(m[1].str())) : 0;
-  const T im = m[6].matched ? static_cast<T>(stold(m[6].str())) : 0;
+  const T re = m[1].matched ? static_cast<T>(stod(m[1].str())) : 0;
+  const T im = m[6].matched ? static_cast<T>(stod(m[6].str())) : 0;
   val = {re, im};
 }
 } // namespace
